FIX subprocess call for Windows
The call made with subprocess.POpen all used shell=False. This works on
Linux and mac, but is an issue on Windows. In this case Windows expects
a full string to the executable. The code only provides a shell command,
and we should set shell=True for Windows.

This has the drawback, that if ever an explicit string command is used,
the logic needs to be refined.